Added prop database as an object so it resets properly on clients
