+++ Assets/Prefabs/Props/PropDatabase.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4f2220205f8542242b2b062d3ea47c2d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
