+++ Assets/Scripts/Prop/PropDatabase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a1725508e1230534b89ba2fc0f38ab77
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
